Refactor report button
